+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Add `ky` frontend dependency as HTTP fetch library
 - Spin the running icon in the status badge
 - Add a live logs indicator to the logs viewer
+- Add duration timer to run page
 
 ### Fixed
 - Sometimes logs are appended to an existing logs files of previous runs (#131)